Refactor: Move Screen sealed class to shared module
This commit moves the `Screen` sealed class, which defines the different screens in the application, from the `navigation` module to the `shared` module.

This change is made to allow other modules, such as `feature:home`, to access the `Screen` definitions.

Additionally, new screen definitions for `ProductsOverview`, `Cart`, and `Categories` have been added to the `Screen` sealed class.

The `navigation` module now depends on the `shared` module to access these definitions.