Refactor: Move Android Compose UI tooling dependency
This commit moves the `androidx.compose.ui.tooling` dependency from the `commonMain` source set to the `androidMain` source set in the following modules:
- `shared`
- `feature/profile`
- `feature/auth`

This change ensures that the UI tooling dependency is only included in the Android-specific build, optimizing the build process for other platforms.

Additionally, the `feature/profile` module has been added as a dependency to the `di` module.