Feat: Add delete product functionality
This commit introduces the ability to delete a product from the `ManageProductScreen`.

A dropdown menu has been added to the top app bar, which appears only when editing an existing product (i.e., `id` is not null). This menu contains a "Delete" option.

Clicking "Delete" triggers the `viewModel.deleteProduct` function. On successful deletion, the user is navigated back. If an error occurs, an error message is displayed using `messageBarState`.