Refactor: Rename FirestoreProductMapper to ProductFirebase
This commit renames the file `FirestoreProductMapper.kt` to `ProductFirebase.kt` within the `com.federico.mapper` package.

This change is purely a renaming of the file and does not involve any logical changes to the code within the file. The purpose is likely to better reflect the content or usage of the class contained within.